+++ lib/functions/cap_data_download.sh
@@ -9,7 +9,11 @@ cap_data_download() {
   # to be removed but we will wait until we have a real example
   # before we try to change this.
   local file_name
-  file_name=$(basename "$cap_data_download_url")
+  if [ -n "$cap_data_download_file_name" ]; then
+    file_name="$cap_data_download_file_name"
+  else
+    file_name=$(basename "$cap_data_download_url")
+  fi
   local output_name
   output_name="${file_name//\.tar.*/}"
   output_name="${output_name//\.gz/}"
@@ -74,20 +78,24 @@ cap_data_download() {
 
 cap_data_download_parse_commandline_parameters() {
   # Define the named commandline options
-  if ! OPTIONS=$(getopt -o "" --long md5sum:,unzip,subdirectory: -- "$@"); then
+  if ! OPTIONS=$(getopt -o "" --long file-name:,md5sum:,unzip,subdirectory: -- "$@"); then
     echo "See CAPTURE help for cap_data_download." >&2
     exit 1
   fi
   eval set -- "$OPTIONS"
 
   # Set default values for the named parameters
+  cap_data_download_file_name=""
   cap_data_download_md5sum=""
   cap_data_download_subdirectory=""
   cap_data_download_unzip=false
-  
+
   # Parse the optional named command line options
   while true; do
     case "$1" in
+      --file-name)
+        cap_data_download_file_name="$2"
+        shift 2 ;;
       --md5sum)
         cap_data_download_md5sum="$2"
         shift 2 ;;
@@ -102,7 +110,7 @@ cap_data_download_parse_commandline_parameters() {
         break;;
     esac
   done
-  
+
   # Check that the required file url parameter was provided
   if [ "$#" -ne 1 ]; then
     echo "Error: incorrect number of parameters" >&2