+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Require the env, run, and verify commands to be ran from the project root
 directory.
 * Rename bin/conda to bin/env
+* Add --file-name option to cap_data_download
 
 ## CAPTURE 0.8.6 (October 21, 2025) ##
 * Fix cap_container failure on Slurm array jobs